pyproject.toml
[tool.poetry]
name = "fast-care"
version = "0.1.0"
description = ""
authors = ["magdy <midomagdy2021@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.3"
psycopg2-binary = "^2.9.3"
pip = "^22.0.4"
install = "^1.3.5"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
